dsxa_sdk/pyproject.toml
[project]
name = "dsxa-sdk"
version = "0.1.3"
description = "Python SDK for Deep Instinct DSX Application Scanner (DSXA) REST APIs"
authors = [
    {name = "Deep Instinct"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.2",
    "pydantic>=2.6.1",
    "typer>=0.12.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
dsxa = "dsxa_sdk.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["dsxa_sdk*"]
